-Added Token model and record into DB (TokenStatus, Token, TokenType, TokenRepo) -Added apache commons libs for method random generate refeshToken (generateRefreshToken method)

+++ jwt-auth/src/main/java/com/jwt/auth/repository/PrivilegeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("PrivilegeRepository")
-public interface PrivilegeRepository extends CrudRepository<Privilege, Integer> {
+public interface PrivilegeRepository extends CrudRepository<Privilege, Long> {
 
 
 }
